Upgrade the _poudriere_freebsd_version
The problems with the previous implementation were:
 * hardcoded versions of FreeBSD, which would easily bitrot
 * no support for SVN branches

The new implementation creates a list of FreeBSD versions based on two
things: freebsd-version(1) and /usr/share/misc/bsd-family-tree. It is
safe to assume that both of them exist on a machine running an
interactive Bash session.

Thanks to freebsd-version(1) the function knows the version of FreeBSD
running on the host machine. As a result it is possible to exclude
FreeBSD versions, which are no longer relevant (most likely have already
reached EoL). The list of currently supported branches is based on the
contents of the /usr/share/misc/bsd-family-tree file, which should give
pretty accurate results.